Functioning DashboardCollectionViewController
