+++ schulcloud/Base.lproj/TabDashboard.storyboard
@@ -558,10 +558,10 @@
         <scene sceneID="dqA-aP-ZXy">
             <objects>
                 <collectionViewController id="Bik-lY-S0X" customClass="DashboardCollectionViewController" customModule="schulcloud" customModuleProvider="target" sceneMemberID="viewController">
-                    <collectionView key="view" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" layoutMarginsFollowReadableWidth="YES" showsHorizontalScrollIndicator="NO" dataMode="prototypes" id="ODr-AP-Vf1">
+                    <collectionView key="view" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" showsHorizontalScrollIndicator="NO" dataMode="prototypes" id="ODr-AP-Vf1">
                         <rect key="frame" x="0.0" y="0.0" width="768" height="1024"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                        <color key="backgroundColor" red="0.60633540149999998" green="0.78882354499999996" blue="0.82049781079999995" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                         <viewLayoutGuide key="safeArea" id="ZOG-nw-4Ox"/>
                         <inset key="scrollIndicatorInsets" minX="5" minY="0.0" maxX="5" maxY="0.0"/>
                         <collectionViewLayout key="collectionViewLayout" id="tBD-4Y-paX" customClass="DashboardLayout" customModule="schulcloud" customModuleProvider="target"/>
